pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opaux"
version = "0.1.0"
description = "Automated job application pipeline powered by Claude AI"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Gordon Akaminko" }]
requires-python = ">=3.11"
keywords = ["jobs", "cv", "resume", "automation", "ai", "claude"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
]

dependencies = [
    "anthropic>=0.40.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "rich-click>=1.7.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "python-docx>=1.1.0",
    "playwright>=1.40.0",
    "python-jobspy>=1.1.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-pyyaml",
    "types-requests",
]
web = [
    "flask>=3.0.0",
    "flask-login>=0.6.0",
    "flask-wtf>=1.2.0",
    "flask-limiter>=3.5.0",
    "flask-mail>=0.10.0",
    "werkzeug>=3.0.0",
    "stripe>=9.0.0",
    "apscheduler>=3.10.0",
    "gunicorn>=22.0.0",
]
pdf = [
    "docx2pdf>=0.1.8",
    "pdfplumber>=0.11.0",
]

[project.scripts]
opaux = "main:cli"

[project.urls]
Homepage = "https://github.com/Gordon-Akaminko/opaux"
Issues = "https://github.com/Gordon-Akaminko/opaux/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
warn_unused_ignores = true
warn_return_any = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
